% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.outliers.R
\name{detect.outliers}
\alias{detect.outliers}
\title{Detect outliers}
\usage{
detect.outliers(data)
}
\arguments{
\item{data}{A matrix or data frame of FPKM values, organized with transcripts on rows and samples on columns.  Transcript identifiers should be stored as \code{rownames(data)}.}
}
\description{
Detect outliers
}
