% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.outliers.R
\name{detect.outliers.sans.null.data}
\alias{detect.outliers.sans.null.data}
\title{Detect outliers, without \code{null.data}}
\usage{
detect.outliers.sans.null.data(
  data,
  num.null,
  p.value.threshold = 0.05,
  kmeans.nstart = 1
)
}
\arguments{
\item{data}{A matrix or data frame of FPKM values, organized with transcripts on rows and samples on columns.  Transcript identifiers should be stored as \code{rownames(data)}.}

\item{num.null}{The number of transcripts to generate when simulating from null distributions.}

\item{p.value.threshold}{The p-value threshold for the outlier test; default is 0.05.  Once the p-value for a sample exceeds \code{p.value.threshold}, testing for that transcript ceases, and all samples (including the one that initially exceeded \code{p.value.threshold}) are assigned p-values equal to \code{p.value.threshold}.}

\item{kmeans.nstart}{The number of random starts when computing k-means fraction; default is 1.  See \code{?stats::kmeans} for further details.}
}
\value{
A list consisting of the following entries:
\itemize{
\item \code{p.values}: a matrix of p-values for the outlier test run on each sample for each transcript in \code{data}.
\item \code{num.outliers}: a vector giving the number of outliers (specifically, the number of samples for which the outlier test yielded a p-value less than \code{p.value.threshold})  for each transcript.
\item \code{outlier.statistics.matrix.list}: a list of length \code{max(num.outliers) + 1} containing entries \code{outlier.statistics.matrix.N}, where \code{N} is between zero and \code{max(num.outliers)}.  \code{outlier.statistics.matrix.N} is the matrix of outlier statistics after excluding the \code{N}th outlier sample, with \code{outlier.statistics.matrix.0} being for the full data set.  A transcript will only appear in the matrices up to the total number of outliers in that transcript; all transcripts appear in \code{outlier.statistics.matrix.0}.
\item \code{distributions}: a numeric vector indicating the optimal distribution for each transcript.  Possible values are 1 (normal), 2 (log-normal), 3 (exponential), and 4 (gamma).
}
}
\description{
Detect outliers, without \code{null.data}
}
