% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.for.calculate.p.values.R
\docType{data}
\name{example.data.for.calculate.p.values}
\alias{example.data.for.calculate.p.values}
\title{example.data.for.calculate.p.values}
\format{
An object of class \code{list} of length 13.
}
\usage{
example.data.for.calculate.p.values
}
\description{
Example data (list object) for testing \code{calculate.p.values()}.
}
\keyword{datasets}
