% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outliers}
\alias{outliers}
\title{Example data set for outlier testing}
\format{
A data frame with 500 rows and 50 columns:
\describe{
\item{S01}{simulated fragments per kilobase of transcript per million fragments mapped (FPKM) values for sample 1}
\item{S02}{simulated FPKM values for sample 2}
\item{S03}{simulated FPKM values for sample 3}
\item{S04}{simulated FPKM values for sample 4}
\item{S05}{simulated FPKM values for sample 5}
\item{S06}{simulated FPKM values for sample 6}
\item{S07}{simulated FPKM values for sample 7}
\item{S08}{simulated FPKM values for sample 8}
\item{S09}{simulated FPKM values for sample 9}
\item{S10}{simulated FPKM values for sample 10}
\item{S11}{simulated FPKM values for sample 11}
\item{S12}{simulated FPKM values for sample 12}
\item{S13}{simulated FPKM values for sample 13}
\item{S14}{simulated FPKM values for sample 14}
\item{S15}{simulated FPKM values for sample 15}
\item{S16}{simulated FPKM values for sample 16}
\item{S17}{simulated FPKM values for sample 17}
\item{S18}{simulated FPKM values for sample 18}
\item{S19}{simulated FPKM values for sample 19}
\item{S20}{simulated FPKM values for sample 20}
\item{S21}{simulated FPKM values for sample 21}
\item{S22}{simulated FPKM values for sample 22}
\item{S23}{simulated FPKM values for sample 23}
\item{S24}{simulated FPKM values for sample 24}
\item{S25}{simulated FPKM values for sample 25}
\item{S26}{simulated FPKM values for sample 26}
\item{S27}{simulated FPKM values for sample 27}
\item{S28}{simulated FPKM values for sample 28}
\item{S29}{simulated FPKM values for sample 29}
\item{S30}{simulated FPKM values for sample 30}
\item{S31}{simulated FPKM values for sample 31}
\item{S32}{simulated FPKM values for sample 32}
\item{S33}{simulated FPKM values for sample 33}
\item{S34}{simulated FPKM values for sample 34}
\item{S35}{simulated FPKM values for sample 35}
\item{S36}{simulated FPKM values for sample 36}
\item{S37}{simulated FPKM values for sample 37}
\item{S38}{simulated FPKM values for sample 38}
\item{S39}{simulated FPKM values for sample 39}
\item{S40}{simulated FPKM values for sample 40}
\item{S41}{simulated FPKM values for sample 41}
\item{S42}{simulated FPKM values for sample 42}
\item{S43}{simulated FPKM values for sample 43}
\item{S44}{simulated FPKM values for sample 44}
\item{S45}{simulated FPKM values for sample 45}
\item{S46}{simulated FPKM values for sample 46}
\item{S47}{simulated FPKM values for sample 47}
\item{S48}{simulated FPKM values for sample 48}
\item{S49}{simulated FPKM values for sample 49}
\item{S50}{simulated FPKM values for sample 50}
}
}
\usage{
outliers
}
\description{
Example data set for outlier testing
}
\keyword{datasets}
