% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zrange.R
\name{zrange}
\alias{zrange}
\title{Range of z-scores}
\usage{
zrange(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A number.
}
\description{
Compute the range of a vector of z-scores.
}
\examples{
set.seed(1234);
x <- rnorm(
    n = 10
    );
zrange(
    x = x
    );
}
